Spotify Plugin v1.0.0
Added rate limiting to not hammer the spotify api, limit is currently set at max 180 requests per 1 minute.

User need to add their own API from the Spotify developer dashboard to the ini file.

A refresh token will be generated so that the user dont have to log in and approve on spotify every time the plugin runs.