Spotify cover art view beta
Added cover art view for spotify.

SpotifyPlugin fetches track cover via url and saves the image as spotifyCoverArt.png in the same folder as the config ini file.

Had to modify LockedImages to be able to read image file changes. Whenever there is an update to the image, it updates in the UI.

Currently the file path for the spotifyCoverArt.png file has to be hardcoded into the profile xml file.

Simply add the full path to the image in the xml under the displayitem like this:

<FilePath>C:\Users\USERNAME\AppData\Local\InfoPanel\plugins\InfoPanel.Extras\spotifyCoverArt.png</FilePath>

Next steps will be do make the image work without hardcoding path in xml, and optimize code. the memory usage is about 50-60mb higher